refactor project structire
