+++ src/App.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { Provider } from 'react-redux';
 import { PersistGate } from 'redux-persist/integration/react';
-import { persistor, store } from './config/store';
 import Navigation from './navigation';
+import { persistor, store } from './redux/store';
 
 const App = () => (
   <Provider store={store}>
