--- src/modules/DecksList/index.ts
@@ -1,3 +1,0 @@
-import DecksList from './components/DecksList';
-
-export { DecksList };
