+++ src/redux/seclectors.ts
@@ -1,5 +1,5 @@
 import { createSelector } from 'reselect';
-import { RootState } from '../../../config/store';
+import { RootState } from './store';
 
 export const selectAllDecks = (state: RootState) => state.decks;
 export const selectDecks = createSelector([selectAllDecks], (decks) => decks);
