+++ src/redux/store.ts
@@ -3,7 +3,7 @@ import { createLogger } from 'redux-logger';
 import { persistStore, persistReducer } from 'redux-persist';
 import AsyncStorage from '@react-native-community/async-storage';
 import autoMergeLevel2 from 'redux-persist/lib/stateReconciler/autoMergeLevel2';
-import decks, { DecksState } from '../modules/DecksList/redux/reducer';
+import decks, { DecksState } from './reducer';
 
 const middleware: Middleware[] = [];
