--- src/screens/Home.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-import { DecksList } from 'modules/DecksList';
-import { Container } from '../common';
-
-const Home = () => (
-  <Container>
-    <DecksList />
-  </Container>
-);
-
-export default Home;
