+++ src/screens/Home/DecksList/index.tsx
@@ -5,9 +5,9 @@ import * as R from 'ramda';
 import { SharedElement } from 'react-navigation-shared-element';
 import DeckItem from './DeckItem';
 import Button from '../../../common/Button';
-import useDecks from '../hooks/useDecks';
 import { Screens } from '../../../navigation/interface';
 import { moderateScale, SPACING, WINDOW_HEIGHT } from '../../../styles/utils';
+import useDecks from '../../../hooks/useDecks';
 
 const colors = ['#fc9d9a', '#f9cdad', '#c8c8a9', '#83af9b', '#d6e1c7', '#94c7b6'];
