+++ src/screens/Home/index.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import { Container } from '../../common';
+import DecksList from './DecksList';
+
+const Index = () => (
+  <Container>
+    <DecksList />
+  </Container>
+);
+
+export default Index;
