+++ src/screens/QuestionModal.tsx
@@ -4,10 +4,10 @@ import { RouteProp } from '@react-navigation/native';
 import { StackNavigationProp } from '@react-navigation/stack';
 import { useDispatch, useSelector } from 'react-redux';
 import { RootStackParamList, Screens } from '../navigation/interface';
-import { saveQuestion } from '../modules/DecksList/redux/actions';
-import { selectCard } from '../modules/DecksList/redux/seclectors';
-import { Card } from '../modules/DecksList/redux/reducer';
 import { Form } from 'common';
+import { selectCard } from '../redux/seclectors';
+import { Card } from '../redux/reducer';
+import { saveQuestion } from '../redux/actions';
 
 type AddQuestionScreenRouteProp = RouteProp<RootStackParamList, Screens.QUESTION_MODAL>;
 type AddQuestionScreenNavigationProp = StackNavigationProp<RootStackParamList, Screens.QUESTION_MODAL>;
